HeadlessGameServerConsole - Break up the class into two, add a unit test, and a performance improvement to add a loop delay in a fast spinning loop.
The controller functionality of HeadlessGameServerConsole was split off into a new class. The console class can now focus on reading input and sending those values to the controller.
